Fixing complaint by MSVC compiler about ambiguous call to overloaded function sqrt

+++ Couenne/src/heuristics/CouenneFPcreateMILP.cpp
@@ -280,7 +280,7 @@ void ComputeSquareRoot (const CouenneFeasPump *fp,
   // Add distance part
   for (int i=0; i<n; ++i)
     if (fp -> Problem () -> Var (i) -> Multiplicity () > 0)
-      A [i * (n+1)] += fp -> multDistMILP () / sqrt (n);
+       A [i * (n+1)] += fp -> multDistMILP () / sqrt (static_cast<double>(n));
 
   // Add gradient-parallel term to the hessian, (x_z - x_z_0)^2. This
   // amounts to setting the diagonal element to GRADIENT_WEIGHT. Don't
